It works

+++ Projekt.c
@@ -107,7 +107,7 @@ char *ReadPixels(int f, int *NumCh)
 	}
 	lseek(f, zsemle, SEEK_SET);
 	read(f, pixel, kenyer * 3);
-	*NumCh = kenyer;
+	*NumCh = (int)kenyer;
 	free(q);
 	free(p);
 	return pixel;
